Make tayga buildable with gcc -Wextra (#131)
Used @Skyb0rg007's suggestion for (void)arg

* tayga.c: address Wsign-compare

* nat64.c: address Wsign-compare

The original litech.org source had often used `int` to store buffer
lengths for some unknown reason, paving the way for possible negative
offset bugs. We fix them all to use `uint32_t`, since the lengths we
deal with fit in 32 bits nicely, they are usually upper bounded by outer
IPv6 packet length.

cfg->ipv6_offlink_mtu could be a `uint16_t`, I decided not to fiddle
with struct field offsets and alignment for now.

* tayga.h: make MTU_ADJ explicitly unsigned

* xlate_header_6to4: address Wunused-parameter

Fixes: 8578d24f9669 ("Documentation & Work (#59)")

* conffile.c: address Wunused-parameter

We mark those as "maybe-unused" with a compiler attribute.

* Review changes

---------

+++ tayga.c
@@ -363,12 +363,12 @@ static void read_from_tun(void)
 				"device: %s\n", strerror(errno));
 		return;
 	}
-	if (ret < sizeof(struct tun_pi)) {
+	if ((size_t)ret < sizeof(struct tun_pi)) {
 		slog(LOG_WARNING, "short read from tun device "
 				"(%d bytes)\n", ret);
 		return;
 	}
-	if (ret == gcfg->recv_buf_size) {
+	if ((uint32_t)ret == gcfg->recv_buf_size) {
 		slog(LOG_WARNING, "dropping oversized packet\n");
 		return;
 	}
@@ -691,7 +691,7 @@ int main(int argc, char **argv)
 
 	if (pidfile) {
 		snprintf(addrbuf, sizeof(addrbuf), "%ld\n", (long)getpid());
-		if (write(pidfd, addrbuf, strlen(addrbuf)) != strlen(addrbuf)) {
+		if (write(pidfd, addrbuf, strlen(addrbuf)) != (ssize_t)strlen(addrbuf)) {
 			slog(LOG_CRIT, "Error, unable to write PID file.\n");
 			exit(1);
 		}
